Create head.c
A program that prints the first 10 lines of its input. If a filename is provided on the command line then it should open it, read and print the first 10 lines, and finally close it. If no filename is provided, head should read from standard input.
.
.
.
don't know if we need this but nice to have for extra points...